services/ads-api/requirements.txt
fastapi
pydantic
python-dotenv
motor
python-jose
cloudinary
uvicorn
python-multipart